# [SofaSparseSolver] SparseLDLSolver optimizations