--- a/modules/SofaSparseSolver/src/SofaSparseSolver/SparseLDLSolver.h
+++ b/modules/SofaSparseSolver/src/SofaSparseSolver/SparseLDLSolver.h
@@ -19,6 +19,11 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
+/******************************************************************************
+* Contributors:
+*   - jeremie.allard@insimo.fr (InSimo)
+*******************************************************************************/
+
 #ifndef SOFA_COMPONENT_LINEARSOLVER_SPARSELDLSOLVER_H
 #define SOFA_COMPONENT_LINEARSOLVER_SPARSELDLSOLVER_H
 #include <SofaSparseSolver/config.h>