… examples/Benchmark/Performance/TorusFall.scn
… examples/Benchmark/Performance/torus1-16x8.obj