@fredroy 2021-04-27
More precise results for each optimisation:
```
no cache + double: 5000 iterations done in 199.849 s ( 25.0188 FPS).
no cache + 3x3   : 5000 iterations done in 170.294 s ( 29.3609 FPS).
cache + double   : 5000 iterations done in 108.942 s ( 45.8959 FPS).
cache + 3x3      : 5000 iterations done in 79.4698 s ( 62.917 FPS).
```