@jnbrunet 2021-04-27
Thanks a lot @fredroy .

Can you share your test scene?

Also, what happen if the default 3x3 block sparse matrix is used with something else than 3D DOFs? For example, with rigids?