@fredroy 2021-06-09
We still did not test with Rigid or Vec2/6.
In my opinion we should be conservative and just let the caching optimization (and let the default template as it is for the moment)
What do you think ?