@jnbrunet 2021-06-25
> what's your opinion @jnbrunet ?

Isn't a warning when the scalar template is used a bit intense? I would put a warning if we actually checked that all mechanical objects are of 3d types, and the template is a scalar. But here, we output a warning without any logic other than the default scalar template was used. Maybe at least remove the warning if the user manually specified that he needed the scalar template?